refactor(ui): Move HoverCard into Block component
Simplify component structure by integrating hover card functionality directly
into Block component rather than wrapping blocks with HoverCard in multiple
places. This reduces code duplication and improves reusability.

Changes:
- Add hoverContent prop to Block component
- Move HoverCardContent to Block.tsx
- Update usages in properties-panel to use new Block API
